Add error display for video-interview analysis and landmark formation, and update FacialFeatureDetector class

+++ modules/main/controllers/DefaultController.php
@@ -115,6 +115,8 @@ class DefaultController extends Controller
                     $videoResultFiles = array();
                     $jsonResultFiles = array();
                     $questions = array();
+                    // Массив для хранения сообщений об ошибках
+                    $errorMassages = array();
                     // Выборка всех цифровых масок у данного видео-интервью
                     $landmarks = Landmark::find()->where(['video_interview_id' => $videoInterviewModel->id])->all();
                     // Обход по всем найденным цифровым маскам
@@ -154,48 +156,50 @@ class DefaultController extends Controller
                     fwrite($jsonFile, str_replace("\\", "", $jsonParameters));
                     // Закрытие файла
                     fclose($jsonFile);
-                    // Статус обработки видеоинтервью
-                    $success = false;
                     // Запуск программы обработки видео
                     chdir($mainPath);
                     exec('./venv/bin/python ./main.py ./test.json');
-                    // Отлов ошибки выполнения программы обработки видео
-                    try {
-                        $index = 0;
-                        $analysisResultIds = '';
-                        $lastAnalysisResultId = null;
-                        // Обход по всем найденным цифровым маскам
-                        foreach ($landmarks as $landmark) {
-                            // Получение значения текста вопроса
-                            $questionText = Yii::$app->request->post('Landmark')[$index]['questionText'];
-                            // Если поле текста вопроса содержит значение "hidden"
-                            if ($questionText != 'hidden') {
-                                // Создание и сохранение новой модели вопроса
-                                $questionModel = new Question();
-                                $questionModel->text = $questionText;
-                                $questionModel->save();
-                                // Формирование id вопроса
-                                $landmark->question_id = $questionModel->id;
-                            }
-                            // Увеличение индекса на 1
-                            $index++;
-                            // Формирование названия json-файла с результатами обработки видео
-                            $landmark->landmark_file_name = 'out_' . $landmark->id . '.json';
-                            // Формирование описания цифровой маски
-                            $landmark->description = $videoInterviewModel->description . ' (время нарезки: ' .
-                                $landmark->getStartTime() . ' - ' . $landmark->getFinishTime() . ')';
-                            // Формирование значения поворота
-                            $landmark->rotation = $rotation;
-                            // Формирование значения наличия отзеркаливания
-                            $landmark->mirroring = boolval($mirroring);
-                            // Обновление атрибутов цифровой маски в БД
-                            $landmark->updateAttributes(['landmark_file_name', 'description', 'rotation',
-                                'mirroring', 'question_id']);
-                            // Проверка существования json-файл с результатами обработки видео
-                            if (file_exists($jsonResultPath . $landmark->landmark_file_name)) {
-                                // Получение json-файла с результатами обработки видео в виде цифровой маски
-                                $landmarkFile = file_get_contents($jsonResultPath .
-                                    $landmark->landmark_file_name, true);
+                    $index = 0;
+                    $analysisResultIds = '';
+                    $lastAnalysisResultId = null;
+                    // Обход по всем найденным цифровым маскам
+                    foreach ($landmarks as $landmark) {
+                        // Получение значения текста вопроса
+                        $questionText = Yii::$app->request->post('Landmark')[$index]['questionText'];
+                        // Если поле текста вопроса содержит значение "hidden"
+                        if ($questionText != 'hidden') {
+                            // Создание и сохранение новой модели вопроса
+                            $questionModel = new Question();
+                            $questionModel->text = $questionText;
+                            $questionModel->save();
+                            // Формирование id вопроса
+                            $landmark->question_id = $questionModel->id;
+                        }
+                        // Увеличение индекса на 1
+                        $index++;
+                        // Формирование названия json-файла с результатами обработки видео
+                        $landmark->landmark_file_name = 'out_' . $landmark->id . '.json';
+                        // Формирование описания цифровой маски
+                        $landmark->description = $videoInterviewModel->description . ' (время нарезки: ' .
+                            $landmark->getStartTime() . ' - ' . $landmark->getFinishTime() . ')';
+                        // Формирование значения поворота
+                        $landmark->rotation = $rotation;
+                        // Формирование значения наличия отзеркаливания
+                        $landmark->mirroring = boolval($mirroring);
+                        // Обновление атрибутов цифровой маски в БД
+                        $landmark->updateAttributes(['landmark_file_name', 'description', 'rotation',
+                            'mirroring', 'question_id']);
+                        $success = false;
+                        // Проверка существования json-файл с результатами обработки видео
+                        if (file_exists($jsonResultPath . $landmark->landmark_file_name)) {
+                            // Получение json-файла с результатами обработки видео в виде цифровой маски
+                            $landmarkFile = file_get_contents($jsonResultPath .
+                                $landmark->landmark_file_name, true);
+                            // Декодирование json-файла с результатами обработки видео в виде цифровой маски
+                            $jsonLandmarkFile = json_decode($landmarkFile, true);
+                            // Если в json-файле с цифровой маской нет текста с ошибкой
+                            if (!isset($jsonLandmarkFile['err_msg'])) {
+                                $success = true;
                                 // Сохранение файла с лицевыми точками на Object Storage
                                 $osConnector->saveFileToObjectStorage(
                                     OSConnector::OBJECT_STORAGE_LANDMARK_BUCKET,
@@ -246,10 +250,11 @@ class DefaultController extends Controller
                                         if (property_exists($frameData, $targetPropertyName) === True)
                                             foreach ($initialData as $frameIndex => $frameData) {
                                                 $actionUnits = $frameData->{$targetPropertyName};
-                                                $actionUnitsAsFacts = $facialFeatureDetector->convertActionUnitsToFacts(
-                                                    $actionUnits,
-                                                    $frameIndex
-                                                );
+                                                $actionUnitsAsFacts = $facialFeatureDetector
+                                                    ->convertActionUnitsToFacts(
+                                                        $actionUnits,
+                                                        $frameIndex
+                                                    );
                                                 if (isset($facts[$frameIndex]) && count($actionUnitsAsFacts) > 0)
                                                     $facts[$frameIndex] = array_merge($facts[$frameIndex],
                                                         $actionUnitsAsFacts);
@@ -272,33 +277,30 @@ class DefaultController extends Controller
                                 // Запоминание последнего id анализа результата
                                 $lastAnalysisResultId = $analysisResultModel->id;
                             } else
-                                // Удаление записи о цифровой маски для которой не сформирован json-файл
-                                Landmark::findOne($landmark->id)->delete();
+                                // Добавление в массив ошибок сообщения об ошибке
+                                array_push($errorMassages, $jsonLandmarkFile['err_msg']);
                         }
-                        // Если есть результаты определения признаков
-                        if ($analysisResultIds != '') {
-                            // Формирование параметров запуска модуля интерпретации признаков
-                            $parameters = array('DataSource' => 'ExecuteReasoningForSetOfInitialConditions',
-                                'AddressForCodeOfKnowledgeBaseRetrieval' =>
-                                    'http://84.201.129.65/default/knowledge-base-download',
-                                'AddressForInitialConditionsRetrieval' =>
-                                    'http://84.201.129.65/analysis-result/facts-download/',
-                                'IDsOfInitialConditions' => '[' . $analysisResultIds . ']',
-                                'AddressToSendResults' => 'http://84.201.129.65:9999/Drools/RetrieveData.php');
-                            // Вызов модуля интерпретации признаков через CURL
-                            $request = curl_init('http://84.201.129.65:9999/Drools/RetrieveData.php');
-                            $dataToSend = http_build_query($parameters);
-                            curl_setopt($request, CURLOPT_POSTFIELDS, $dataToSend);
-                            curl_setopt($request, CURLOPT_RETURNTRANSFER, True);
-                            curl_exec($request);
-                            curl_close($request);
-                        }
-                        // Изменение статуса обработки видеоинтервью
-                        $success = true;
-                    } catch (Exception $e) {
-                        // Вывод сообщения об ошибке обработки видеоинтервью
-                        Yii::$app->getSession()->setFlash('error',
-                            'При обработке видеоинтервью возникли ошибки!');
+                        if ($success == false)
+                            // Удаление записи о цифровой маски для которой не сформирован json-файл
+                            Landmark::findOne($landmark->id)->delete();
+                    }
+                    // Если есть результаты определения признаков
+                    if ($analysisResultIds != '') {
+                        // Формирование параметров запуска модуля интерпретации признаков
+                        $parameters = array('DataSource' => 'ExecuteReasoningForSetOfInitialConditions',
+                            'AddressForCodeOfKnowledgeBaseRetrieval' =>
+                                'http://84.201.129.65/default/knowledge-base-download',
+                            'AddressForInitialConditionsRetrieval' =>
+                                'http://84.201.129.65/analysis-result/facts-download/',
+                            'IDsOfInitialConditions' => '[' . $analysisResultIds . ']',
+                            'AddressToSendResults' => 'http://84.201.129.65:9999/Drools/RetrieveData.php');
+                        // Вызов модуля интерпретации признаков через CURL
+                        $request = curl_init('http://84.201.129.65:9999/Drools/RetrieveData.php');
+                        $dataToSend = http_build_query($parameters);
+                        curl_setopt($request, CURLOPT_POSTFIELDS, $dataToSend);
+                        curl_setopt($request, CURLOPT_RETURNTRANSFER, True);
+                        curl_exec($request);
+                        curl_close($request);
                     }
                     // Удаление файла с видеоинтервью
                     if (file_exists($videoPath . $videoInterviewModel->video_file_name))
@@ -313,22 +315,47 @@ class DefaultController extends Controller
                     foreach ($jsonResultFiles as $jsonResultFile)
                         if (file_exists($jsonResultPath . $jsonResultFile))
                             unlink($jsonResultPath . $jsonResultFile);
-                    // Если видеоинтервью обработалось корректно
-                    if ($success) {
-                        // Если был сформирован результат анализа
-                        if ($lastAnalysisResultId != null) {
-                            // Вывод сообщения об успешном анализе видеоинтервью
-                            Yii::$app->getSession()->setFlash('success',
-                                'Вы успешно проанализировали видеоинтервью!');
+                    // Если был сформирован результат анализа
+                    if ($lastAnalysisResultId != null) {
+                        // Вывод сообщения об успешном анализе видеоинтервью
+                        Yii::$app->getSession()->setFlash('success',
+                            'Вы успешно проанализировали видеоинтервью!');
 
-                            return $this->redirect(['/analysis-result/view/' . $lastAnalysisResultId]);
-                        } else {
-                            // Вывод сообщения о неуспешном анализе видеоинтервью
-                            Yii::$app->getSession()->setFlash('error',
-                                'Видеоинтервью проанализировать не удалось!');
-
-                            return $this->redirect(['/video-interview/view/' . $videoInterviewModel->id]);
+                        return $this->redirect(['/analysis-result/view/' . $lastAnalysisResultId]);
+                    } else {
+                        // Текст сообщения об ошибке
+                        $errorMessage = 'Не удалось проанализировать видеоинтервью!';
+                        // Проверка существования json-файл с ошибками обработки видеоинтервью в корневой папке
+                        if (file_exists($mainPath . 'error.json')) {
+                            // Получение json-файл с ошибками обработки видеоинтервью
+                            $jsonFile = file_get_contents($mainPath . 'error.json', true);
+                            // Декодирование json
+                            $jsonFile = json_decode($jsonFile, true);
+                            // Дополнение текста сообщения об ошибке
+                            $errorMessage .= PHP_EOL . $jsonFile['err_msg'];
+                            // Удаление json-файла с сообщением ошибки
+                            unlink($mainPath . 'error.json');
                         }
+                        // Проверка существования json-файл с ошибками обработки видеоинтервью в папке json
+                        if (file_exists($jsonResultPath . 'out_error.json')) {
+                            // Получение json-файл с ошибками обработки видеоинтервью
+                            $jsonFile = file_get_contents($jsonResultPath . 'out_error.json',
+                                true);
+                            // Декодирование json
+                            $jsonFile = json_decode($jsonFile, true);
+                            // Дополнение текста сообщения об ошибке
+                            $errorMessage .= PHP_EOL . $jsonFile['err_msg'];
+                            // Удаление json-файла с сообщением ошибки
+                            unlink($jsonResultPath . 'out_error.json');
+                        }
+                        // Дополнение текста сообщения об ошибке - ошибками по отдельным вопросам
+                        if (!empty($errorMassages))
+                            foreach ($errorMassages as $errorMassage)
+                                $errorMessage .= PHP_EOL . $errorMassage;
+                        // Вывод сообщения о неуспешном формировании цифровой маски
+                        Yii::$app->getSession()->setFlash('error', $errorMessage);
+
+                        return $this->redirect(['/video-interview/view/' . $videoInterviewModel->id]);
                     }
                 }
             }
